strip out old homepage
